# docs/swagger.yaml
basePath: /
definitions:
  api.GetSentMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/api.SentMessageResponse'
        type: array
      total:
        type: integer
    type: object
  api.MessageRequest:
    properties:
      content:
        type: string
      to:
        type: string
    required:
    - content
    - to
    type: object
  api.MessageResponse:
    properties:
      message:
        example: accepted
        type: string
      messageId:
        example: 01623bff-7fa9-4ccb-a843-e6d98908dc49
        type: string
    type: object
  api.SentMessageResponse:
    properties:
      content:
        example: Hello, World!
        type: string
      createdAt:
        example: "2023-10-01T10:00:00Z"
        type: string
      id:
        example: 01623bff-7fa9-4ccb-a843-e6d98908dc49
        type: string
      phone:
        example: "+905551234567"
        type: string
      remoteMessageId:
        example: whatsapp-msg-123
        type: string
      sentAt:
        example: "2023-10-01T10:05:00Z"
        type: string
      status:
        example: SENT
        type: string
      updatedAt:
        example: "2023-10-01T10:05:00Z"
        type: string
    type: object
  api.StopSchedulerResponse:
    properties:
      message:
        example: Message scheduler stopped successfully
        type: string
      status:
        example: success
        type: string
    type: object
  server.HealthResponse:
    properties:
      message:
        example: Message scheduler alive!
        type: string
      status:
        example: ok
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a message scheduler service
  title: Message Scheduler API
  version: "1.0"
paths:
  /_monitoring/health:
    get:
      description: Check if the message scheduler service is running
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            $ref: '#/definitions/server.HealthResponse'
      summary: Health Check
      tags:
      - monitoring
  /sent-messages:
    get:
      description: Retrieve sent messages with optional limit
      parameters:
      - description: 'Number of messages to retrieve (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sent messages retrieved successfully
          schema:
            $ref: '#/definitions/api.GetSentMessagesResponse'
        "400":
          description: Invalid limit parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve sent messages
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Sent Messages
      tags:
      - messages
  /start-send-message:
    post:
      consumes:
      - application/json
      description: Send a message via webhook and return the webhook response
      parameters:
      - description: Message request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message sent successfully
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Invalid request body or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to send message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start Send Message
      tags:
      - messages
  /stop-message-sender:
    post:
      description: Stop the currently running message-sending scheduler
      produces:
      - application/json
      responses:
        "200":
          description: Scheduler stopped successfully
          schema:
            $ref: '#/definitions/api.StopSchedulerResponse'
        "500":
          description: Failed to stop scheduler
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stop Message Sender
      tags:
      - messages
swagger: "2.0"